Update __main__.py
Hi, 
I want to use your script as base to continue developing more features, but for now I am just testing if I can use it or not. If you have any suggestions is welcome as I know this has not been edited for a while.

I am trying to test the script to do some further works. I was facing this issue on python 3.9.2. 

"
Traceback (most recent call last):
  File "/app/defichain_exporter/defichain_exporter/__main__.py", line 62, in <module>
    main()
  File "/app/defichain_exporter/defichain_exporter/__main__.py", line 44, in main
    start_http_server(args.port)
(...)
TypeError: an integer is required (got type str)"

Requires an int in newer versions of python, but the parser passes it as string.

This could also be done by parsing integer on the argparse side before passing it to the server, but like this is good enough for now. Please consider approving the change.